LayoutTests/fast/frames/srcdoc/srcdoc-defaults-to-standards-mode.html

<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();
</script>
Even through the srcdoc below doesn't have a DocType, it should still be in
standards mode.
<iframe srcdoc="
  <script>
  alert(document.compatMode);
  </script>
"></iframe>
