third_party/blink/web_tests/fast/frames/srcdoc/srcdoc-defaults-to-standards-mode-expected.txt

ALERT: CSS1Compat
Even through the srcdoc below doesn't have a DocType, it should still be in standards mode. 
